Clean up & menubar addition
