+++ adventure.datetime/res/menu/authordesc.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android" >
+
+    <item
+        android:id="@+id/action_edit"
+        android:icon="@drawable/ic_action_edit"
+        android:orderInCategory="1"
+        android:showAsAction="always"
+        android:title="Edit"/>
+    
+    <item 
+        android:id="@+id/action_upload"
+        android:icon="@drawable/ic_action_upload"
+        android:orderInCategory="2"
+        android:showAsAction="always"
+        android:title="Upload"/>
+    
+    <item
+        android:id="@+id/action_discard"
+        android:icon="@drawable/ic_action_edit"
+        android:orderInCategory="3"
+        android:showAsAction="always"
+        android:title="Discard"/>
+</menu>
